requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
pandas>=2.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
urllib3>=2.0.0
pytest>=7.4.0
